+++ src/main/java/com/kuit/conet/dao/TeamDao.java
@@ -1,8 +1,9 @@
+/*
 package com.kuit.conet.dao;
 
 import com.kuit.conet.domain.team.Team;
 import com.kuit.conet.domain.team.TeamMember;
-import com.kuit.conet.dto.web.response.StorageImgResponseDTO;
+import com.kuit.conet.dto.web.response.member.StorageImgResponseDTO;
 import com.kuit.conet.dto.web.response.team.GetTeamMemberResponseDTO;
 import com.kuit.conet.dto.web.response.team.GetTeamResponseDTO;
 import lombok.extern.slf4j.Slf4j;
@@ -321,3 +322,4 @@ public class TeamDao {
         return jdbcTemplate.query(sql, param, mapper);
     }
 }
+*/
